+++ src/test/kotlin/features/glue/StepDefs.kt
@@ -28,7 +28,7 @@ import kotlin.test.fail
  * @date 2019-04-28
  */
 class StepDefs {
-    @Given("^a (snakemake|snakemake:5x|snakemake:6.1|snakemake with disabled framework|python) project$")
+    @Given("^a (snakemake|snakemake:5x|snakemake:6.1|snakemake:6.5|snakemake with disabled framework|python) project$")
     @Throws(Exception::class)
     fun configureSnakemakeProject(projectType: String) {
         require(SnakemakeWorld.myFixture == null) {